[WIP] 3.8 core

+++ tests/Bridge/Bridge_ApiTest.php
@@ -15,28 +15,26 @@ class Bridge_ApiTest extends PhraseanetPHPUnitAuthenticatedAbstract
     public function setUp()
     {
         parent::setUp();
-        $appbox = appbox::get_instance(\bootstrap::getCore());
 
         $sql = 'DELETE FROM bridge_apis WHERE name = "Apitest"';
-        $stmt = $appbox->get_connection()->prepare($sql);
+        $stmt = self::$DI['app']['phraseanet.appbox']->get_connection()->prepare($sql);
         $stmt->execute();
         $stmt->closeCursor();
 
         $this->type = 'Apitest';
-        $api = Bridge_Api::create($appbox, $this->type);
+        $api = Bridge_Api::create(self::$DI['app'], $this->type);
 
         $this->id = $api->get_id();
-        $this->object = new Bridge_Api($appbox, $api->get_id());
+        $this->object = new Bridge_Api(self::$DI['app'], $api->get_id());
     }
 
     public function tearDown()
     {
-        $appbox = appbox::get_instance(\bootstrap::getCore());
-
-        $this->object->delete();
-
+        if ($this->object) {
+            $this->object->delete();
+        }
         try {
-            new Bridge_Api($appbox, $this->id);
+            new Bridge_Api(self::$DI['app'], $this->id);
             $this->fail();
         } catch (Bridge_Exception_ApiNotFound $e) {
